ref(demo): Move span tree demo guide
